Create the News Print stories and render them
