+++ src/components/NewsPrintStories.js
@@ -1,0 +1,14 @@
+import React from 'react';
+import NewsTitleText from './NewsTitleText';
+
+function NewsPrintStories() {
+  return (
+    <NewsTitleText
+      titleText="PRINT STORIES"
+      topStoriesTitle="Lorem Ipsum Dolor."
+      storiesBody="Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor <br> incididunt ut labore."
+    />
+  );
+}
+
+export default NewsPrintStories;
